+++ Assets/OxGFrame/AssetLoader/Scripts/Editor/Bundle/EditorWindow/BundleUrlConfigGeneratorWindow.cs
@@ -64,7 +64,7 @@ namespace OxGFrame.AssetLoader.Editor
 
             GUIStyle style = new GUIStyle();
             var bg = new Texture2D(1, 1);
-            ColorUtility.TryParseHtmlString("#1c589c", out Color color);
+            ColorUtility.TryParseHtmlString("#1e3836", out Color color);
             Color[] pixels = Enumerable.Repeat(color, Screen.width * Screen.height).ToArray();
             bg.SetPixels(pixels);
             bg.Apply();